requirements.txt
# Core dependencies for Search and Deep Research Agents
python-dotenv>=1.0.0
azure-identity>=1.15.0
aiohttp>=3.9.0

# OpenAI API for chat completion and function calling
openai>=1.35.0

# Additional Azure dependencies (optional)
azure-ai-projects>=1.0.0
azure-core>=1.29.0